move terraform stuff into a separate file
